requirements.txt
twilio
flask
requests
openai
python-dotenv
